requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
feedparser==6.0.11
httpx==0.28.1
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
apscheduler==3.10.4
rapidfuzz==3.10.1
lxml==5.3.0
psycopg2-binary==2.9.10
google-genai
groq>=1.0.0
python-dotenv==1.0.1
python-jose[cryptography]
authlib
itsdangerous
resend
